Branch explicitly for node.js and Express implementation of server code. Connected to accounts database for login, but no real functionality (yet). It's after 3 am.
