--- logs/requestLog.txt
@@ -1,100 +1,0 @@
-20241102	00:22:40	0192eb1c-6912-72b4-9e18-12c15b7fe81e	/	GET
-20241102	00:22:40	0192eb1c-691b-7717-8cae-781305b76bcd	/css/styles.css	GET
-20241102	00:22:40	0192eb1c-691c-715c-8a7f-8033837823a5	/imgs/basic%20logo.jpeg	GET
-20241102	00:22:40	0192eb1c-691d-72f1-9623-b55edb47c5b6	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:22:40	0192eb1c-6921-74f8-b095-c0ff8081775a	/scripts/index.js	GET
-20241102	00:22:40	0192eb1c-693a-762b-b12a-e357a8cc56c5	/favicon.ico	GET
-20241102	00:25:06	0192eb1e-a41f-73ad-b3fc-cfe16e02b2f9	/	GET
-20241102	00:25:06	0192eb1e-a426-76af-b37b-d71b64af4cfb	/css/styles.css	GET
-20241102	00:25:06	0192eb1e-a428-732f-9a92-aca9f65e1aa4	/imgs/basic%20logo.jpeg	GET
-20241102	00:25:06	0192eb1e-a429-71fe-b83b-12664f5e5c2b	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:25:06	0192eb1e-a438-7548-9b9d-bb4bd1854322	/scripts/index.js	GET
-20241102	00:25:06	0192eb1e-a44b-726b-8e19-b74b9beb3107	/favicon.ico	GET
-20241102	00:26:25	0192eb1f-d654-76dc-b85c-b0f717aae34d	/player.html	GET
-20241102	00:26:25	0192eb1f-d658-75a6-91b3-e9f402fadc80	/css/styles.css	GET
-20241102	00:26:25	0192eb1f-d659-7168-933e-bcb20dcde3b7	/scripts/player.js	GET
-20241102	00:26:25	0192eb1f-d659-7168-933e-c2058840ada8	/imgs/basic%20logo.jpeg	GET
-20241102	00:26:25	0192eb1f-d666-75e8-a5f0-f48693c0fe64	/videos/PHP%20local%20server.mp4	GET
-20241102	00:26:25	0192eb1f-d672-72a9-9714-7009804ab18c	/favicon.ico	GET
-20241102	00:26:28	0192eb1f-e26c-7047-aabb-4e4cf1fd613f	/favicon.ico	GET
-20241102	00:29:13	0192eb22-67f9-7773-97e0-021f86cb33d5	/	GET
-20241102	00:29:13	0192eb22-6801-75fe-86eb-2a067f37f517	/css/styles.css	GET
-20241102	00:29:13	0192eb22-6801-75fe-86eb-2ebc1d595cfd	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:29:13	0192eb22-6802-7332-bec8-dd259d8719a4	/imgs/basic%20logo.jpeg	GET
-20241102	00:29:13	0192eb22-6803-743d-b812-b15517044a70	/scripts/index.js	GET
-20241102	00:29:13	0192eb22-680d-75ae-9895-aea3deef8670	/favicon.ico	GET
-20241102	00:29:18	0192eb22-7ad2-73ca-978e-bdb57c536b83	/index.html	GET
-20241102	00:29:18	0192eb22-7ad6-7357-9b28-e3603d060c6d	/css/styles.css	GET
-20241102	00:29:18	0192eb22-7ad7-743e-b4f7-d88bc383e31e	/imgs/basic%20logo.jpeg	GET
-20241102	00:29:18	0192eb22-7ad7-743e-b4f7-de92454d87bd	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:29:18	0192eb22-7ad8-762f-b88e-e3c81f01bf79	/scripts/index.js	GET
-20241102	00:31:17	0192eb24-4d6e-77bd-897f-2391c2e35bb6	/index.html	GET
-20241102	00:31:17	0192eb24-4d76-758f-b3d9-04d4254f9820	/css/styles.css	GET
-20241102	00:31:17	0192eb24-4d77-72d6-8496-d9f0220f131d	/imgs/basic%20logo.jpeg	GET
-20241102	00:31:17	0192eb24-4d78-702b-a85c-adb635feeef7	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:31:17	0192eb24-4d78-702b-a85c-b35593447307	/scripts/index.js	GET
-20241102	00:31:56	0192eb24-e29d-75dc-bcd1-6db3df8d145b	/index.html	GET
-20241102	00:31:56	0192eb24-e2a1-708c-b7f0-6e8aa58e8bb5	/css/styles.css	GET
-20241102	00:31:56	0192eb24-e2a3-71c9-acda-060866b7f89f	/imgs/basic%20logo.jpeg	GET
-20241102	00:31:56	0192eb24-e2a3-71c9-acda-08294d106056	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:31:56	0192eb24-e2a3-71c9-acda-0fd7f858d858	/scripts/index.js	GET
-20241102	00:32:37	0192eb25-83b6-7199-bf5b-625a2245e27a	/index.html	GET
-20241102	00:32:37	0192eb25-83bd-7018-916b-af235f461801	/css/styles.css	GET
-20241102	00:32:37	0192eb25-83be-777e-80a4-24b72ee3988a	/imgs/basic%20logo.jpeg	GET
-20241102	00:32:37	0192eb25-83bf-7039-a58c-e34a52b532ff	/scripts/index.js	GET
-20241102	00:32:37	0192eb25-83be-777e-80a4-28d7bd9a66ef	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:32:38	0192eb25-88de-7261-887b-f71e05aeda5a	/index.html	GET
-20241102	00:32:38	0192eb25-88e2-72dc-9af8-9e4fbca0c7e3	/css/styles.css	GET
-20241102	00:32:38	0192eb25-88e2-72dc-9af8-a285e03ec7da	/imgs/basic%20logo.jpeg	GET
-20241102	00:32:38	0192eb25-88e3-729a-abee-53213407cafb	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:32:38	0192eb25-88e3-729a-abee-573a5e32bda4	/scripts/index.js	GET
-20241102	00:32:39	0192eb25-8c72-75ea-ae60-c28059c49c8a	/upload.html	GET
-20241102	00:32:39	0192eb25-8c77-761e-9c08-f889144f8488	/css/styles.css	GET
-20241102	00:32:39	0192eb25-8c78-7609-8784-c801197b52cb	/imgs/basic%20logo.jpeg	GET
-20241102	00:32:39	0192eb25-8c79-73d3-bcce-5d3c117b5a37	/scripts/upload.js	GET
-20241102	00:32:41	0192eb25-92ca-70ac-a6d5-0509a3f4cb59	/index.html	GET
-20241102	00:32:41	0192eb25-92ce-72da-86bd-20aacf530e76	/css/styles.css	GET
-20241102	00:32:41	0192eb25-92cf-7189-a151-7416e9057529	/imgs/basic%20logo.jpeg	GET
-20241102	00:32:41	0192eb25-92cf-7189-a151-7b218b55287a	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:32:41	0192eb25-92d0-7678-ac1a-982b59c2f761	/scripts/index.js	GET
-20241102	00:32:42	0192eb25-9698-718a-820d-820295be12f6	/login.php	GET
-20241102	00:32:42	0192eb25-969e-754c-81ab-47ffb9cd58f2	/scripts/login.js	GET
-20241102	00:32:42	0192eb25-969e-754c-81ab-4aa96282236f	/css/styles.css	GET
-20241102	00:32:42	0192eb25-969f-717c-b9e8-f541aef46df0	/imgs/basic%20logo.jpeg	GET
-20241102	00:34:41	0192eb27-686e-737f-85a9-fb5d3de2f5d7	/	GET
-20241102	00:34:41	0192eb27-6879-730c-a22b-52b42ffed47e	/css/styles.css	GET
-20241102	00:34:41	0192eb27-687b-759b-9737-b8a41ea6f0dd	/imgs/basic%20logo.jpeg	GET
-20241102	00:34:41	0192eb27-687c-766e-a2fa-3f85869ce513	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:34:41	0192eb27-687f-758a-9b65-f486ab05cc4a	/scripts/index.js	GET
-20241102	00:38:08	0192eb2a-91d2-75cd-a729-64c3259e6c02	/	GET
-20241102	00:38:08	0192eb2a-91d8-71ca-badf-646e365deb3c	/css/styles.css	GET
-20241102	00:38:08	0192eb2a-91d9-7354-be19-50f94f18d86a	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:38:08	0192eb2a-91da-7148-bb79-fbaacc4dcec8	/imgs/basic%20logo.jpeg	GET
-20241102	00:38:08	0192eb2a-91db-75be-a6fa-6720908b65fd	/scripts/index.js	GET
-20241102	00:38:50	0192eb2b-35b0-705e-aa2c-eb1df8321eec	/player.html	GET
-20241102	00:38:50	0192eb2b-35b5-75f9-a6cf-0a9100754551	/css/styles.css	GET
-20241102	00:38:50	0192eb2b-35b5-75f9-a6cf-0d4003a5b465	/imgs/basic%20logo.jpeg	GET
-20241102	00:38:50	0192eb2b-35b6-735f-b575-7359dafeb8c7	/scripts/player.js	GET
-20241102	00:38:50	0192eb2b-35c2-72d8-ad81-eae324e1da86	/videos/PHP%20local%20server.mp4	GET
-20241102	00:39:29	0192eb2b-cef4-703c-a402-71500a649ce3	/	GET
-20241102	00:39:29	0192eb2b-ceff-747e-aff8-1184a1af4728	/css/styles.css	GET
-20241102	00:39:29	0192eb2b-cf00-76bf-8544-9e67003fcb81	/imgs/basic%20logo.jpeg	GET
-20241102	00:39:29	0192eb2b-cf01-7469-8a57-d0589fed721a	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:39:29	0192eb2b-cf02-7769-9b27-ac851c373fc7	/scripts/index.js	GET
-20241102	00:41:22	0192eb2d-885d-76c1-9bbe-2575a00e679c	/player.html	GET
-20241102	00:41:22	0192eb2d-8865-762d-b475-f4b623a4dba4	/css/styles.css	GET
-20241102	00:41:22	0192eb2d-8866-72e8-b558-1c834a5df771	/imgs/basic%20logo.jpeg	GET
-20241102	00:41:22	0192eb2d-8866-72e8-b558-201fc6b4550c	/scripts/player.js	GET
-20241102	00:41:22	0192eb2d-8873-77fc-888b-ab55a082000b	/videos/PHP%20local%20server.mp4	GET
-20241102	00:41:25	0192eb2d-91b6-7472-8431-78bbfc2407da	/videos/PHP%20local%20server.mp4	GET
-20241102	00:43:52	0192eb2f-d0a9-7107-af51-7991080d7655	/	GET
-20241102	00:43:52	0192eb2f-d0b5-70c6-8505-28fa1eb19bd1	/css/styles.css	GET
-20241102	00:43:52	0192eb2f-d0b7-7085-a05c-39de79158854	/imgs/basic%20logo.jpeg	GET
-20241102	00:43:52	0192eb2f-d0b7-7085-a05c-3f26e3394800	/imgs/PHP%20code%20thumbnail.png	GET
-20241102	00:43:52	0192eb2f-d0c7-7108-b55a-d1d1b07bba90	/scripts/index.js	GET
-20241102	00:43:54	0192eb2f-da05-77b4-9692-3dc3a7da765c	/player.html	GET
-20241102	00:43:54	0192eb2f-da0a-70ad-ac50-f38025c38a41	/css/styles.css	GET
-20241102	00:43:54	0192eb2f-da0b-735b-8de4-eebe938a6b82	/imgs/basic%20logo.jpeg	GET
-20241102	00:43:54	0192eb2f-da0c-74bb-846d-5a6865062ba1	/scripts/player.js	GET
-20241102	00:43:54	0192eb2f-da1c-71cd-bd6f-4c9c18cf579e	/videos/PHP%20local%20server.mp4	GET
-20241102	00:43:54	0192eb2f-dae0-720b-b7e1-398eea2b23d6	/videos/PHP%20local%20server.mp4	GET
