+++ package.json
@@ -22,7 +22,10 @@
   },
   "homepage": "https://github.com/SmoCloud/video-player#readme",
   "dependencies": {
+    "cors": "^2.8.5",
     "date-fns": "^4.1.0",
+    "express": "^4.21.1",
+    "mysql": "^2.18.1",
     "uuid": "^11.0.2"
   },
   "devDependencies": {
