+++ public/scripts/player.js
