+++ views/index.html
@@ -3,20 +3,22 @@
     <head>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="../css/styles.css">
+        <link rel="stylesheet" href="css/styles.css">
         <title>Z^2</title>
     </head>
     <body>
         <section class="border">
             <div class="navbar">
-                <img src="../imgs/basic logo.jpeg">
-                <a href="login.php">Login</a>
+                <img src="imgs/basic logo.jpeg">
+                <a href="login.html">Login</a>
                 <a href="upload.html">Upload</a>
                 <a href="index.html">Home</a>
             </div>
             <div class="searchbar">
-                <input type="text" id="videoTitle" style="width: 60%;">
-                <button id="searchButton" style="border-radius: 5px;">Search</button>
+                <form action="index.html" method="post">
+                    <input type="text" id="videoTitle" style="width: 60%;">
+                    <button id="searchButton" style="border-radius: 5px;">Search</button>
+                </form>
             </div>
         </section>
         <section class="recommendations">
@@ -25,7 +27,7 @@
                 <hr color="lightgrey">
                 <div class="vDemo_1">
                     <label for="fruitninja" style="font-family: Arial; font-size: 1.5em;">Fruit Ninja Demo</label>
-                    <hr color="black"><img id="fruitninja" src="../imgs/PHP code thumbnail.png" style="width: 100%; border-radius: 16px;">
+                    <hr color="black"><img id="fruitninja" src="imgs/PHP code thumbnail.png" style="width: 100%; border-radius: 16px;">
                 </div>
                 <div class="vDemo_2">
 
@@ -41,6 +43,6 @@
                     
             </div>
         </section>
-        <script src="../scripts/index.js"></script>
+        <script src="scripts/index.js"></script>
     </body>
 </html>
