+++ views/login.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="stylesheet" href="css/styles.css">
+        <title>Login</title>
+    </head>
+    <body>
+        <section class="border">
+            <div class="navbar">
+                <img src="imgs/basic logo.jpeg">
+                <a href="login.html">Login</a>
+                <a href="index.html">Home</a>
+            </div>
+            <h1 style="text-align: center;">Login/Create Account</h1>
+        </section>
+        <form class="login-box" action="login.html" method="post">
+            <div class="username">
+                <label for="usr">Username: </label>
+                <input type="text" id="usr" name="usr" style="width: 50%;">
+            </div>
+            <div class="password">
+                <label for="pwd">Password: </label>
+                <input type="password" id="pwd" name="pwd" style="width: 50%;">
+            </div>
+            <div class="account-btns">
+                <button id="login-btn" name="login" style="border-radius: 5px;" value="Login">Login</button>
+                <button id="create-acct-btn" name="create" style="border-radius: 5px;" value="Create Account">Create Account</button>
+            </div>
+        </form>
+        <script src="scripts/login.js"></script>
+    </body>
+</html>
