+++ views/upload.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Upload Page</title>
-    <link rel="stylesheet" href="../css/styles.css">
+    <link rel="stylesheet" href="css/styles.css">
 </head>
 <body>
     <div class="navbar">
-        <img src="../imgs/basic logo.jpeg">
+        <img src="imgs/basic logo.jpeg">
         <a href="index.html">Home</a>
         <a href="upload.html">Upload</a>
     </div>
@@ -17,12 +17,12 @@
         <button id="searchButton" style="border-radius: 5px;">Search</button>
     </div>
     <div class="uploading">
-    	<form action="upload.php" method="post" enctype="multipart/form-data">
+    	<form action="upload" method="post" enctype="multipart/form-data">
 		  Select video to upload:
 		  <input type="file" name="fileToUpload" id="fileToUpload">
 		  <input type="submit" value="Upload Video" name="submit">
 		</form>
     </div>
-    <script src="../scripts/upload.js"></script>
+    <script src="scripts/upload.js"></script>
 </body>
 </html>
